==> R/01_Demography_data_preparation.Rmd <==
---
title: "PVA of the NBI: Demography - Data Preparation" ## name of your project
author: "Sinah Drenske"                     ## your name(s)
date: "`r Sys.Date()`"                  ## current date
output:
  rmdformats::readthedown:
    code_folding: hide                  ## hide or show code by default?
    toc_depth: 3                        ## 3-level TOC
---

```{r setup, include=FALSE}
## You can ignore this chunk in most cases
## If you want to modify chunk options, you can do it here for all chunks or
## add the options in the repsective chunk header, e.g. `{r, message = FALSE}`
knitr::opts_chunk$set(echo = TRUE, warning = TRUE, message = TRUE,
                      fig.width = 9, fig.height = 6, dpi = 500, 
                      retina = 1, fig.showtext = TRUE)
```

* **Hypothesis:** We hypothesize that at the present time the northern bald ibis population can survive without further management and release. We predict that the observed demographic rates will ensure population growth and do not differ between the breeding colonies. 
* **Study area:** Austria, Germany and Italy
* **Data:** Life History data and Breeding data of the NBI

In this script we prepare the data for analysis. We format columns, add new columns or change the order of the columns.

# Setup
```{r packages}
## libraries used in this script
library("here")
library("tidyverse")
library("eeptools")
```


# Data
## Load data
```{r data}
NBI <- read_csv2(file = here::here("data-raw", "NBI_life_history_data_2008_2019.csv"), 
                 col_names = T, # first row as column headers
                 col_types = "dcffDffccccffDDff", # specify column types
                 n_max = 411) # There are 411 NBI

NBI_breeding <- read_delim(file = here::here("data-raw", "NBI_breeding_data_2008_2019.csv"), 
                           delim =  ";", 
                           escape_double = FALSE, 
                           col_types = cols(b_pair_date = col_date(format = "%d.%m.%Y"), 
                                            b_parent_f_id = col_number(), 
                                            b_parent_m_id = col_number(), 
                                            b_breeding_colony = col_factor(levels = c("Burghausen", "Kuchl")), 
                                            b_count_eggs = col_number(), 
                                            b_count_hatching = col_number(), 
                                            b_count_fledgling = col_number()), 
                           na = "NA", 
                           trim_ws = TRUE) 
```

# NBI
## Prepare data
### Format columns
We have to format the column "sex". Individuals where the sex is unknown receive the letter "u" instead of "n.a."

```{r}
table(NBI$sex)

NBI$sex <- ifelse(test = NBI$sex == "n.a.", yes = "u", no = paste(NBI$sex))

table(NBI$sex)
NBI$sex <- as.factor(NBI$sex)
levels(NBI$sex)
```

### Create additional variables
Create columns for hatch year, death month and death year
```{r}
# Hatch year, death month and death year with substring from the hatching or death date
NBI$hatch_year <- substring(NBI$hatching_date, 1,4)
NBI$death_month <- substring(NBI$delivered_or_death_date, 6,7)
NBI$death_year <- substring(NBI$delivered_or_death_date, 1,4)

# If an individual is alive and part of the project put a zero as death year 
NBI$death_year[is.na(NBI$death_year)] <- 0

# Create a variable for the date
# The data is from 2008 to 2019. In order to avoid miscalculating the survival probabilities, all NBI that were alive at the end of 2019 receive the date of death 01.01.2020
day <- as.Date("2020-01-01")
NBI[is.na(NBI$delivered_or_death_date), "delivered_or_death_date"] <- as.Date(day)

# Change one date of death, because there was a mistake (the date of death was before the hatching date). We set the death date to the same date as the hatching date
# row 399 
NBI$delivered_or_death_date[399] <- as.Date("2017-06-09")

# remove the variable day
rm(day)
```

Calculate the age of the NBI: days of life (dof) and years of life (yof)
Pay attention! First year of life = Age 0. Second year of life = Age 1, ...
```{r}
# Use the function age_calc from the eeptools package
# dob = date of birth/start date, enddate = date when the observation's age is of interest, 
# units = days, months or years; precise = include leap years or not/ only full years or not
NBI$dof <- as.numeric(age_calc(dob = NBI$hatching_date, enddate = NBI$delivered_or_death_date, units = "days", precise = F))
# +1 for the years of life because we want the years of life NOT the age
NBI$yof <- (age_calc(dob = NBI$hatching_date, enddate = NBI$delivered_or_death_date, units = "years", precise = F) +1)
```

Add a column for the stage (age class) 
```{r}
# reached_stage
NBI$reached_stage <- ifelse(
                       test = NBI$yof == 1, 
                       yes = 1, 
                       no = ifelse(
                              test = NBI$yof == 2, 
                              yes = 2, 
                              no = ifelse(
                                     test = NBI$yof == 3, 
                                     yes = 3, 
                                     no = ifelse(
                                            test = NBI$yof  > 3, 
                                            yes = 4, 
                                            no = 0))))

NBI$reached_stage <- as.factor(NBI$reached_stage)
```


Add one column for the status: 
Was the individual at the time of the death date alive or dead? This will be a numeric variable. Some NBI were removed from the project although they were still alive. For example, they were too seriously injured. 
0 = alive (status = alive, temporary or delivered), 1 = dead
```{r}
# life_status
NBI$life_status <- ifelse(test = NBI$status == "dead", 
                          yes = 1, 
                          no = 0)
NBI$life_status <- as.factor(NBI$life_status)

# Testing
table(NBI$status)
table(NBI$life_status)
```


Add one column for the project status: 
Is the individual alive and part of the project or not (dead or no longer part of the project). This will be a numeric variable.
0 = alive and part of the project (status = alive or temporary), 1 = not part of the project anymore (delivered or dead)
```{r}
# project_status
NBI$project_status <- ifelse(test = NBI$status == "alive" | NBI$status == "temporary", 
                             yes = 0, 
                             no = 1)
NBI$project_status <- as.factor(NBI$project_status)

# Testing
table(NBI$status)
table(NBI$project_status)
```

Insert "alive" in the mortality type (mortality cause) for individuals which are still alive and insert "delivered" in the mortality type for individuals that are alive but are not part of the project anymore
```{r}
levels(NBI$mortality_cause)
NBI$mortality_cause <- ifelse(test = NBI$status == "alive" | NBI$status == "temporary", 
                              yes = "alive", 
                              no = ifelse(test = NBI$status == "delivered", 
                                          yes = "delivered", 
                                          no = paste(NBI$mortality_cause))) 

NBI$mortality_cause <- as.factor(NBI$mortality_cause)
NBI$mortality_cause <- droplevels(NBI$mortality_cause)
levels(NBI$mortality_cause)
```

## Sort table
Reorder rows and columns
```{r}
names(NBI)
```

```{r}
# reorder rows
NBI %<>%
  arrange(hatching_date) %>%
  as_tibble()
```


```{r}
#reorder columns
NBI %<>% 
  select(bird_id, name, hatching_date, hatch_year, 
         generation, sex, raising_type, breeding_area, mortality_cause, mortality_verified, 
         missed_since, delivered_or_death_date, death_month, death_year, dof, yof, reached_stage, 
         status, life_status, project_status, 
         parent_f, parent_f_id, parent_m, parent_m_id, population) %>%
  as_tibble()
```


## Filter data
Some NBI have to be removed from the dataset for the analysis:
We consider only fledged individuals in our analysis because there is not data about all individuals that died before fledging. But in the table are some individuals that died before fledging. We remove them from the table. These individuals have the mortality cause "pre_fledgling".
And there is one individual that was born in 2007 but we want to consider only individuals that were born in 2008 or later.

```{r}
NBI_all <- NBI

NBI <- subset(x = NBI_all, subset = !hatch_year == "2007" & !mortality_cause == "pre_fledgling")
NBI$mortality_cause <- droplevels(NBI$mortality_cause)
```


# Save data
```{r}
# saveRDS(object = NBI_all, file = here::here("output", "data-proc", "NBI_all_data.rds"))
# saveRDS(object = NBI, file = here::here("output", "data-proc", "NBI_data.rds"))
# 
# saveRDS(object = NBI_breeding, file = here::here("output", "data-proc", "NBI_breeding.rds"))
```


***

<details><summary>Session Info</summary>

```{r sessionInfo}
## DO NOT REMOVE!
## We store the settings of your computer and the current versions of the
## packages used to allow for reproducibility
Sys.time()
#git2r::repository() ## uncomment if you are using GitHub
sessionInfo()
```

</details>
