==> R/03_Study_area.Rmd <==
---
title: "PVA of the NBI: Study area"
author: "Cédric Scherer"           
date: "`r Sys.Date()`"     
output:
  rmdformats::readthedown:
    code_folding: hide     
    toc_depth: 3           
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  echo = TRUE, warning = FALSE, message = FALSE,
  fig.width = 9, fig.height = 6, dev = "ragg_png", dpi = 180, retina = 1
)
```

* **Hypothesis:** We hypothesize that at the present time the northern bald ibis population can survive without further management and release. We predict that the observed demographic rates will ensure population growth and do not differ between the breeding colonies. 
* **Study area:** Austria, Germany and Italy
* **Data:** Elevation data and the breeding and wintering locations of the NBI population

In this script we create the map for the study area.

# Setup

```{r prep}
## packages
library("tidyverse")
library("raster")
library("sf")
library("RStoolbox")
library("ggsn")
library("lwgeom")
library("rnaturalearth")
library("rmapshaper")
library("shadowtext")
library("patchwork")
library("systemfonts")

## set the theme
theme_set(theme_minimal(base_size = 14, base_family = "Open Sans"))
theme_update(
  axis.ticks = element_line(color = "grey10"),
  axis.ticks.length = unit(-.5, "lines"),
  axis.text.x = element_text(size = 16, 
                             margin = margin(t = 3)),
  axis.text.y = element_text(size = 16, 
                             margin = margin(r = 3)),
  panel.border = element_rect(color = "grey10", fill = NA, size = 1), 
  panel.background = element_rect(color = "white", fill = "white"),
  plot.margin = margin(5, 10, 10, 10),
  plot.tag = element_text(size = 24, face = "bold"),
  legend.position = "top",
  legend.text = element_text(face = "bold",
                             size = 20,
                             margin = margin(2, 6, 2, 6))
)
```


# Data

```{r data}
## elevation data
path_dem <- here::here("output", "geo-proc", "df_dem_alps.Rds")
path_hills <- here::here("output", "geo-proc", "df_hills.Rds")
path_slope <- here::here("output", "geo-proc", "df_slope.Rds")

if(!file.exists(path_hills)) {
  dem <- raster(here::here("data-raw", "geo-raw", "dem_3035_crop.tif"))
  dem_alps <- aggregate(dem, fact = 4)
  
  saveRDS(dem_alps, path_dem)
} else {
  dem_alps <- readRDS(path_dem)
}

## breeding and wintering locations
locations <- 
  tibble(
    location = c("B", "U", "K", "R", "L"),
    type = c("Breeding ground", "Breeding ground", "Breeding ground", "Breeding ground", "Wintering ground"),
    label = c("Burghausen\n(Germany)", "Überlingen\n(Germany)", 
              "Kuchl\n(Austria)", "Rosegg\n(Austria)", "Laguna di\nOrbetello\n(Italy)"),
    label_alt = c("Burghausen", "Überlingen", 
              "Kuchl", "Rosegg", "Laguna di\nOrbetello"),
    side = c("r", "l", "r", "r", "r"),
    x = c(4531605, 4258768, 4557635, 4629096, 4419572),
    y = c(2788488, 2739526, 2728565, 2616771, 2148807),
    from_x = c(4531605, 4258768, 4557635, 4629096, NA_real_),
    to_x = c(4419572, 4419572, 4419572, 4419572, NA_real_),
    from_y = c(2788488, 2739526, 2739526, 2616771, NA_real_),
    to_y = c(2148807, 2148807, 2148807, 2148807, NA_real_)
  ) 
```


# Overview Map

```{r overview-map, fig.width = 8.5, fig.height = 10}
sf_world <- 
  rnaturalearth::ne_countries(returnclass = "sf", scale = 10) %>% 
  st_transform(crs = st_crs(dem_alps)) %>% 
  st_buffer(dist = 0) %>% 
  dplyr::select(iso_a2, sovereignt, continent) %>% 
  mutate(area = st_area(.))

sf_inner <- rmapshaper::ms_innerlines(sf_world)

overview <- 
  ggplot() +
    geom_sf(data = sf_world,
            fill = "grey60",
            color = "transparent") +
    geom_sf(data = sf_inner, 
            color = "grey80", 
            size = .3) +
    geom_rect(xmin = 4000600, xmax = 4840000,
              ymin = 2000600, ymax = 2990000,
              color = "#212121",
              fill = NA) +
    geom_sf_text(data = sf_world %>% 
                   filter(iso_a2 %in% c("DE", "IT", "AT")) %>% 
                   group_by(iso_a2) %>% slice(1),
                 aes(label = iso_a2),
                 family = "Open Sans",
                 color = "white",
                 fontface = "bold",
                 size = 5.7,
                 nudge_x = 20000) +
    geom_point(data = locations,
               aes(x, y),
               shape = 21,
               color = "#212121",
               fill = "white",
               size = 3,
               stroke = .8) +
    ggspatial::annotation_scale(location = 'tl',
                                text_family = "Open Sans",
                                text_cex = 1.2) +
    coord_sf(xlim = c(2000000, 6600000), 
             ylim = c(1000000, 5600000)) +
    scale_x_continuous(expand = c(0, 0), breaks = seq(-10, 30, by = 10)) +
    theme(panel.ontop = FALSE,
          panel.grid.major = element_line(color = "grey85", linetype = "dashed", size = .4)) +
    labs(x = NULL, y = NULL)

overview
```


```{r save-overview-map, fig.width = 8.5, fig.height = 10}
## save map
# ggsave(here::here("plots", "01_description", "map_A_europe_countries.pdf"),
#        width = 8.95, height = 10, device = cairo_pdf)
```


# Hillshade map 

```{r hillshade-map, fig.width = 8.5, fig.height = 10}
dem_alps@data@values <- dem_alps@data@values * 10

if(!file.exists(path_hills)) {
  ## hillshade and slopeshade
  slope <- terrain(dem_alps, opt = "slope", unit = "radians")
  aspect <- terrain(dem_alps, opt = "aspect", unit = "radians")
  hill <- hillShade(slope, aspect, 40, 270)
  
  ## ggplot makes me turn the raster into points
  df_hills <- rasterToPoints(hill)
  df_hills <- data.frame(df_hills)
  colnames(df_hills) <- c("lon", "lat", "hills")
  
  ## the key to a pretty map is merging the slope shade with the hillshade
  df_slope <- rasterToPoints(slope)
  df_slope <-  data.frame(df_slope)
  colnames(df_slope) <- c("lon", "lat", "slope")
  df_slope$slope <- 1 - df_slope$slope ## invert the scale so that more slope is darker
  
  saveRDS(df_hills, path_hills)
  saveRDS(df_slope, path_slope)
} else {
  df_hills <- readRDS(path_hills)
  df_slope <- readRDS(path_slope)
}

## naturalearth data
oceans_raw <- 
  rnaturalearth::ne_download(scale = 10, 
                             category = "physical", 
                             type = "ocean",
                             returnclass = "sf")

bbox <- as(extent(3, 17, 38, 51), 'SpatialPolygons')
crs(bbox) <- crs(oceans_raw)

oceans <-
  oceans_raw %>% 
  st_transform(crs = st_crs(bbox)) %>% 
  st_make_valid() %>% 
  st_crop(., bbox) %>% 
  ## transformation leads to inversed oceans - fix it with ms_erase
  rmapshaper::ms_erase(st_as_sf(bbox), .) %>%
  st_transform(crs = st_crs(dem_alps))

lakes <- 
  rnaturalearth::ne_download(scale = 10, 
                             category = "physical", 
                             type = "lakes",
                             returnclass = "sf") %>% 
  st_transform(crs = st_crs(dem_alps)) 

## map
alps <- 
  ggplot() +
    geom_raster(data = df_hills, aes(lon, lat, fill = hills, group = 1)) +
    geom_raster(data = df_slope, aes(lon, lat, fill = slope, group = 2), alpha = .6) +
    geom_sf(data = oceans,
            fill = "#627696",
            color = "#627696") +
    geom_sf(data = lakes,
            fill = "#627696",
            color = "#627696") +
    geom_curve(data = filter(locations, location %in% c("B", "K", "R")),
               aes(x = from_x, xend = to_x,
                   y = from_y, yend = to_y),
               curvature = .12,
               size = .8,
               color = "#212121") +
    geom_curve(data = filter(locations, location == "U"),
               aes(x = from_x, xend = to_x,
                   y = from_y, yend = to_y),
               curvature = -.12,
               size = .8,
               color = "#212121") +
    ## outer ring
    geom_point(data = locations,
               aes(x, y,
                   color = type),
               shape = 21,
               fill = "#fefefe",
               size = 7,
               stroke = .8) +
    ## inner point
    geom_point(data = locations,
               aes(x, y,
                   color = type),
               size = 3) +
    ## location labels
    geom_shadowtext(data = filter(locations, side == "r"),
                    aes(x, y, label = label),
                    nudge_x = 20000,
                    size = 4.7,
                    color = "#fefefe",
                    family = "Open Sans",
                    fontface = "bold",
                    lineheight = .9,
                    hjust = 0,
                    bg.color = "#212121") +
    geom_shadowtext(data = filter(locations, side == "l"),
                    aes(x, y, label = label),
                    nudge_x = -20000,
                    size = 4.7,
                    color = "#fefefe",
                    family = "Open Sans",
                    fontface = "bold",
                    lineheight = .9,
                    hjust = 1,
                    bg.color = "#212121") +
    ggspatial::annotation_scale(location = 'tl',
                                text_family = "Open Sans",
                                text_cex = 1.2) +
    ggspatial::annotation_north_arrow(location = 'bl') +
    scale_color_manual(values = c("#d10173", "#21bffe"), name = NULL) +
    rcartocolor::scale_fill_carto_c(palette = "ag_GrnYl", direction = -1, guide = "none") +
    #rcartocolor::scale_fill_carto_c(palette = "Fall", direction = -1, guide = "none") +
    scale_x_continuous(expand = c(0, 0), limits = c(4000600, 4840000)) +
    scale_y_continuous(expand = c(0, 0), limits = c(2000600, 2990000)) +
    guides(color = guide_legend(nrow = 1, label.position = "right")) +
    theme(legend.position = "top") +
    labs(x = NULL, y = NULL)

alps


alps_alt <- 
  alps +
  guides(color = guide_legend(ncol = 1, label.position = "right")) +
  theme(legend.position = c(1, 1), legend.justification = c(1.0001, 1.01),
        legend.background = element_rect(fill = "white", color = "grey10",  size = .5),
        legend.margin = margin(7, 5, 10, 10))

alps_alt
```

# Combined Maps

```{r panel, fig.width = 17, fig.height = 10}
p <- overview + alps + plot_annotation(tag_level = "A", tag_suffix = ".")

p
```


### Alternative Variant with Inside Legend

```{r panel-alt, fig.width = 17, fig.height = 10}
p_alt <- overview + alps_alt + plot_annotation(tag_level = "A", tag_suffix = ".") 

p_alt
```

```{r save-panel}
## save panel
file <- here::here("plots", "01_description", "map_panel_GrnYl")

#ggsave(plot = p, paste0(file, ".pdf"), width = 17, height = 10, device = cairo_pdf)
ggsave(plot = p, paste0(file, ".png"), width = 17, height = 10, dpi = 750, bg = "white")
ggsave(plot = p, paste0(file, ".svg"), width = 17, height = 10)


file <- here::here("plots", "01_description", "map_panel_GrnYl_alt")

#ggsave(plot = p_alt, paste0(file, ".pdf"), width = 17, height = 10, device = cairo_pdf)
ggsave(plot = p_alt, paste0(file, ".png"), width = 17, height = 10, dpi = 750, bg = "white")
ggsave(plot = p_alt, paste0(file, ".svg"), width = 17, height = 10)
```

# Alternative Single Map Variant

```{r single map, fig.width = 9, fig.height = 10}
country_labs <- 
  tibble(
    country = c("Germany", "Italy", "Austria"),
    long = c(10.5, 12.2, 14.9),
    lat = c(48.6, 43.3, 47.2)
  ) %>% 
  st_as_sf(coords = c("long", "lat"), crs = crs(oceans_raw))

p_single <-
  ggplot() +
    geom_raster(data = df_hills, aes(lon, lat, fill = hills, group = 1)) +
    geom_raster(data = df_slope, aes(lon, lat, fill = slope, group = 2), alpha = .6) +
    geom_sf(data = oceans,
            fill = "#627696",
            color = "#627696") +
    geom_sf(data = lakes,
            fill = "#627696",
            color = "#627696") +
    geom_sf(data = sf_inner, 
            color = "white",
            size = .9) +
    geom_sf_text(data = country_labs,
                 aes(label = country),
                 family = "Open Sans",
                 color = "white",
                 fontface = "bold",
                 size = 9) +
    geom_curve(data = filter(locations, location %in% c("B", "K", "R")),
               aes(x = from_x, xend = to_x,
                   y = from_y, yend = to_y),
               curvature = .12,
               size = 1.3,
               color = "#212121") +
    geom_curve(data = filter(locations, location == "U"),
               aes(x = from_x, xend = to_x,
                   y = from_y, yend = to_y),
               curvature = -.12,
               size = 1.3,
               color = "#212121") +
    ## outer ring
    geom_point(data = locations,
               aes(x, y,
                   color = type),
               shape = 21,
               fill = "#fefefe",
               size = 9,
               stroke = .8) +
    ## inner point
    geom_point(data = locations,
               aes(x, y,
                   color = type),
               size = 3) +
    ## location labels
    geom_shadowtext(data = filter(locations, side == "r"),
                    aes(x, y, label = label_alt),
                    nudge_x = 20000,
                    size = 7.7,
                    color = "#fefefe",
                    family = "Open Sans",
                    fontface = "bold",
                    lineheight = .9,
                    hjust = 0,
                    bg.color = "#212121") +
    geom_shadowtext(data = filter(locations, side == "l"),
                    aes(x, y, label = label_alt),
                    nudge_x = -20000,
                    size = 7.7,
                    color = "#fefefe",
                    family = "Open Sans",
                    fontface = "bold",
                    lineheight = .9,
                    hjust = 1,
                    bg.color = "#212121") +
    ggspatial::annotation_scale(location = 'tl',
                                text_family = "Open Sans",
                                text_cex = 1.6) +
    ggspatial::annotation_north_arrow(location = 'bl') +
    scale_color_manual(values = c("#d10173", "#21bffe"), name = NULL) +
    rcartocolor::scale_fill_carto_c(palette = "ag_GrnYl", direction = -1, guide = "none") +
    scale_x_continuous(expand = c(0, 0), limits = c(4000000, 4840000)) +
    scale_y_continuous(expand = c(0, 0), limits = c(2050000, 3000000)) +
    #guides(color = guide_legend(nrow = 1, label.position = "right")) +
    guides(color = guide_legend(ncol = 1, label.position = "right")) +
    labs(x = NULL, y = NULL) +
    theme(legend.position = c(1, 1), legend.justification = c(1.0001, 1.01),
          legend.background = element_rect(fill = "white", color = "grey10", size = .5),
          legend.margin = margin(7, 5, 10, 10))

p_single
```

```{r save-single-map}
## save panel
file <- here::here("plots", "01_description", "map_single_GrnYl")

#ggsave(plot = p_single, paste0(file, ".pdf"), width = 9, height = 10, device = cairo_pdf)
ggsave(plot = p_single, paste0(file, ".png"), width = 9, height = 10, dpi = 750, bg = "white")
ggsave(plot = p_single, paste0(file, ".svg"), width = 9, height = 10)
```

```{r cvd-check, fig.width = 18, fig.height = 20}
colorBlindness::cvdPlot(p_single)
```


***

<details><summary>Session Info</summary>

```{r sessionInfo}
Sys.time()
git2r::repository()
sessionInfo()
```

</details>
