==> R/06c_PVA_Management_Improvement_plots.Rmd <==
---
title: "PVA of the NBI: Analysis of Management Improvement Scenarios - Plots" ## name of your project
author: "Sinah Drenske"                     ## your name(s)
date: "`r Sys.Date()`"                  ## current date
output:
  rmdformats::readthedown:
    code_folding: hide                  ## hide or show code by default?
    toc_depth: 3                        ## 3-level TOC
---

```{r setup, include=FALSE}
## You can ignore this chunk in most cases
## If you want to modify chunk options, you can do it here for all chunks or
## add the options in the repsective chunk header, e.g. `{r, message = FALSE}`
knitr::opts_chunk$set(echo = TRUE, warning = TRUE, message = TRUE,
                      fig.width = 9, fig.height = 6, dpi = 500, 
                      retina = 1, fig.showtext = TRUE)
```

* **Hypothesis:** We hypothesize that at the present time the northern bald ibis population can survive without further management and release. We predict that the observed demographic rates will ensure population growth and do not differ between the breeding colonies. 
* **Study area:** Austria, Germany and Italy
* **Data:** Results of the first NetLogo Model and the tables created in script 06b

In this script we create barplots to show the distribution of survival and reproduction values between the scenarios and we show the population trajectories of the 14 scenarios of special interest.

# Setup

```{r packages}
## for non-CRAN packages please keep install instruction
## but commented so it is not run each time, e.g.
# devtools::install_github("EcoDynIZW/template")

## libraries used in this script
## please add ALL LIBRARIES NEEDED HERE
## please remove libraries from the list that are not needed anymore 
## at a later stage
library("extrafont")
library("here")
library("ggplot2")
library("viridisLite")
```


# Data

```{r data}
df_Basic_scen <- readRDS(file = here("output", "data-proc", "model-proc", "df_Basic_scen.rds"))
df_lambda_pos <- readRDS(file = here("output", "data-proc", "model-proc", "df_lambda_pos.rds"))
df_mean <- readRDS(file = here("output", "data-proc", "model-proc", "df_mean.rds"))
df_sc_mean <- readRDS(file = here("output", "data-proc", "model-proc", "df_sc_mean.rds"))
df_scen <- readRDS(file = here("output", "data-proc", "model-proc", "df_scen.rds"))
```

# Barplot of scenarios with positive population development
Table the distributions between the different mortalities and reproduction rates
```{r}
table(df_lambda_pos$Mortality_Juveniles)
# 0.20  0.30  0.36 
#  107   102    99          
table(df_lambda_pos$Mortality_Subadults1)
# 0.08  0.19  0.26   
#  107   102    99          
table(df_lambda_pos$Mortality_Subadults2)
# 0.14  0.24  0.31 
#  107   102    99          
table(df_lambda_pos$Mortality_Adults)
# 0.02  0.14  0.22 
#  108   108    92          
table(df_lambda_pos$Repro)
# 0.53  0.58  0.66  1.06  1.41  3.97 
#   71    74    80    81     1     1   
```
Table the distributions of values across all scenarios
```{r}
table(df_sc_mean$Mortality_Juveniles)
# 0.20  0.30  0.36 
#  108   108   110          
table(df_sc_mean$Mortality_Subadults1)
# 0.08  0.19  0.26   
#  108   108   110          
table(df_sc_mean$Mortality_Subadults2)
# 0.14  0.24  0.31 
#  108   108   110          
table(df_sc_mean$Mortality_Adults)
# 0.02  0.14  0.22 
#  108   108   110          
table(df_sc_mean$Repro)
# 0.53  0.58  0.66  1.06  1.41  3.97 
#   81    81    81    81     1     1   
```

## Version 1: Absolute numbers
We create a matrix, but we leave out the the Status quo and All chicks scenarios. Therefore we will not use these reproduction rates and we enter the baseline survival with two values less than described above

```{r}
mort_RR_mat_all <- matrix(data = c(
  108,108,108,NA,
  108,108,108,NA,
  108,108,108,NA,
  108,108,108,NA,
   81, 81, 81, 81, NA, NA
), byrow = F)

mort_RR_mat <- matrix( data = c(
  97,102, 107,NA,       # s1 baseline, 10%, 25% improved
  97,102, 107,NA,       # s2
  97,102, 107,NA,       # s3
  90,108,108,NA,         # s4
  71,74, 80,81, NA, NA  # RR baseline, 10%, 25%, 100%
), byrow = F)
mort_RR_mat
```
Plot
```{r}
par(mar = c(5,5.1,2,2))
barplot(mort_RR_mat_all, beside = T, ylim = c(0,120), space = c(0,-0.3),  las = 1,
        col = "grey", cex.axis = 1.5, cex.lab = 1.5, ylab = "Count", border = NA)
barplot(mort_RR_mat, beside = T, ylim = c(0,120), space = c(0,-0.3),  las = 1,
        col = viridis(4, end = 0.95), cex.axis = 1.5,  border = NA, add = T)
axis(1, at = c(1.2, 4.9,8.6,12.3,16.5), labels = c("","","","",""), cex.axis = 1.5)
axis(1, at = c(1.2, 4.9,8.6,12.3,16.5), labels = c("s1","s2","s3","s4","RR"), cex.axis = 1.5, line = 0.2, lwd = 0)
box()
```

## Version 2: Percentage
Calculate the percent of the used s1-s4 and reproduction values in scenarios with positive population growth
```{r}
mort_RR_mat_perc <- (mort_RR_mat/mort_RR_mat_all) * 100
mort_RR_mat_perc
```
Plot
```{r}
par(mar = c(4,8,1,1))
barplot(mort_RR_mat_perc, beside = T, ylim = c(0,100), space = c(0,-0.3),  las = 1,
        col = viridis(4, end = 0.95), cex.axis = 3, cex.lab = 3, border = NA, 
        ylab = " ")
#axis(1, at = c(1:20), cex.axis = 2.5)
axis(1, at = c(1.1, 5.1,9.1,13.1,17.7), labels = c("","","","",""), cex.axis = 2.5)
axis(1, at = c(1.1, 5.1,9.1,13.1,17.7), labels = c("s1","s2","s3","s4","RR"), cex.axis = 2.5, line = 0.2, lwd = 0)
mtext("% of used values in scenarios", side = 2, line = 6, cex = 3)
mtext("with positive population growth", side = 2, line = 4, cex = 3)
box()
text(x = -2.8, y = -5, labels = "a)", xpd=NA, cex = 3)
```

Save the plot
```{r}
# png(filename = here("plots", "04_PVA", "BaselineImprovements_bar.png"), width = 1200, height = 772)
# par(mar = c(4,8,1,1))
# barplot(mort_RR_mat_perc, beside = T, ylim = c(0,100), space = c(0,-0.3),  las = 1,
#         col = viridis(4, end = 0.95), cex.axis = 3, cex.lab = 3, border = NA,
#         ylab = " ")
# #axis(1, at = c(1:20), cex.axis = 2.5)
# axis(1, at = c(1.1, 5.1,9.1,13.1,17.7), labels = c("","","","",""), cex.axis = 2.5)
# axis(1, at = c(1.1, 5.1,9.1,13.1,17.7), labels = c("s1","s2","s3","s4","RR"), cex.axis = 2.5, line = 0.2, lwd = 0)
# mtext("% of used values in scenarios", side = 2, line = 6, cex = 3)
# mtext("with positive population growth", side = 2, line = 4, cex = 3)
# box()
# text(x = -2.8, y = -5, labels = "a)", xpd=NA, cex = 3)
# dev.off()
```


# Population trajectories
A plot of the population trajectories of the 14 scnarios of special interest
It will be a spaghetti plot with facets per scenario to show all runs per scenario and the mean of all runs per scenario

```{r}
length(unique(df_Basic_scen$Scenario))
levels(df_Basic_scen$Scenario)

length(unique(df_mean$Scenario))
levels(df_scen$Scenario)
```

Change the order of the factor levels for df_Basic_scen and df_scen
```{r}
df_Basic_scen$Scenario <- factor(
  x = df_Basic_scen$Scenario, 
  levels = c( "0.360.260.310.220.53true04true00.25", "0.360.260.310.220.58true04true00.25", 
              "0.360.260.310.220.66true04true00.25", "0.360.260.310.221.06true04true00.25", 
              "0.30.260.310.220.53true04true00.25", "0.20.260.310.220.53true04true00.25", 
              "0.360.260.310.140.53true04true00.25", "0.360.260.310.020.53true04true00.25", 
              "0.30.190.240.140.53true04true00.25", "0.20.080.140.020.53true04true00.25", 
              "0.30.190.240.140.58true04true00.25", "0.20.080.140.020.66true04true00.25", 
              "0.360.260.310.221.41true04true00.25", "0.360.260.310.223.97true04true00.25"))

df_scen$Scenario <- factor(
  x = df_scen$Scenario, 
  levels = c( "0.360.260.310.220.53true04true00.25", "0.360.260.310.220.58true04true00.25", 
              "0.360.260.310.220.66true04true00.25", "0.360.260.310.221.06true04true00.25", 
              "0.30.260.310.220.53true04true00.25", "0.20.260.310.220.53true04true00.25", 
              "0.360.260.310.140.53true04true00.25", "0.360.260.310.020.53true04true00.25", 
              "0.30.190.240.140.53true04true00.25", "0.20.080.140.020.53true04true00.25", 
              "0.30.190.240.140.58true04true00.25", "0.20.080.140.020.66true04true00.25", 
              "0.360.260.310.221.41true04true00.25", "0.360.260.310.223.97true04true00.25"))
```

```{r}
scenario_names <- c("Baseline","+10% RR", 
                    "+25% RR", "+100% RR", 
                    "+10% s1","+25% s1", 
                    "+10% s4","+25% s4",
                    "+10% s1-s4", "+25% s1-s4", 
                    "+10% s1-s4 & +10% RR", "+25% s1-s4 & +25% RR", 
                    "Status quo", "All chicks") 
names(scenario_names) <- c("0.360.260.310.220.53true04true00.25", "0.360.260.310.220.58true04true00.25", 
              "0.360.260.310.220.66true04true00.25", "0.360.260.310.221.06true04true00.25", 
              "0.30.260.310.220.53true04true00.25", "0.20.260.310.220.53true04true00.25", 
              "0.360.260.310.140.53true04true00.25", "0.360.260.310.020.53true04true00.25", 
              "0.30.190.240.140.53true04true00.25", "0.20.080.140.020.53true04true00.25", 
              "0.30.190.240.140.58true04true00.25", "0.20.080.140.020.66true04true00.25", 
              "0.360.260.310.221.41true04true00.25", "0.360.260.310.223.97true04true00.25")

# df_scen: add a year and a Number for Ind_All for plotting. Will be the same for all rows
df_scen$IndAll <- 150
df_scen$Year <- 1
```

## Plot
Considered scenarios:
+ Baseline

+ 10% reproduction
+ 25% reproduction
+ 100% reproduction

+ 10% s1
+ 25% s1

+ 10% s4
+ 25% s4

+ 10% s1-s4
+ 25% s1-s4

+ 10% s1-s4+ 10% reproduction
+ 25% s1-s4+ 25% reproduction

+ Status quo

+ All chicks
= 14 scenarios

```{r}
length(unique(df_Basic_scen$Run)) # 1400 = 100 runs for 14 scenarios
```
```{r}
length(unique(df_Basic_scen$Scenario))
```


```{r}
# Loading fonts (with package extrafont)
font_import()
loadfonts(device = "win")
fonts()

# Version 1 y-lim 0-200
ggplot(df_Basic_scen, aes(x = Year, y = IndAll)) + 
  geom_line(aes(group = Run, colour = Scenario)) + 
  facet_wrap( ~Scenario, 
              nrow = 7, ncol = 2, 
              labeller = labeller(Scenario = scenario_names)) + 
  geom_line(data = df_mean, aes(x = Year, y = IndAll, group = Scenario), size = 1) +
  scale_y_continuous(limits = c(0, 200)) +
  theme_light() +
  theme(legend.position = "none", 
        panel.grid = element_blank(), 
        strip.background =element_rect(fill="grey50"), 
        text = element_text(family = "Open Sans")) + 
  labs(x = "Years after simulation start", y = "Female individuals") + 
  geom_text(data = df_scen, mapping = aes(x = 42, y = 170),  
            label = paste(sprintf("%.2f", round(df_scen$Lambda_sto, digits = 2)), # sprintf ensures that two (or three) decimal places are actually displayed.
                          "\u00b1", # +- sign
                          sprintf("%.3f", round(df_scen$SD_Lambda, digits = 3)), sep = " "), 
            size = 3) + 
  scale_colour_viridis_d(option = "plasma", begin = 0.4)


# Version 2 y-lim 0-500
# ggplot(df_Basic_scen, aes(x = Year, y = IndAll)) + 
#   geom_line(aes(group = Run, colour = Scenario)) + 
#   facet_wrap( ~Scenario, 
#               nrow = 7, ncol = 2, 
#               labeller = labeller(Scenario = scenario_names)) + 
#   geom_line(data = df_mean, aes(x = Year, y = IndAll, group = Scenario), size = 1) +
#   scale_y_continuous(limits = c(0, 500)) +
#   theme_light() +
#   theme(legend.position = "none", panel.grid = element_blank(), 
#         strip.background =element_rect(fill="grey50")) + 
#   labs(x = "Years after simulation start", y = "Female individuals") + 
#   geom_text(data = df_scen, mapping = aes(x = 45, y = 170),  
#             label = paste(sprintf("%.2f", round(df_scen$Lambda_sto, digits = 2)), # sprintf ensures that two decimal places are actually displayed.
#                           "\u00b1", # +- sign
#                           sprintf("%.2f", round(df_scen$SD_Lambda, digits = 2)), sep = " "), 
#             size = 3) + 
#   scale_colour_viridis_d(option = "plasma", begin = 0.4)
```

Save the plot
```{r}
 # ggsave(filename = here("plots", "04_PVA", "20220801_pop_trajectories_200.png"),
 #        width = 11.5, height = 17.17, units = "cm", dpi = 2800)
 # ggsave(filename = here("plots", "04_PVA", "20220801_pop_trajectories_200.svg"),
 #        width = 11.5, height = 17.17, units = "cm", dpi = 2800)
```

***

<details><summary>Session Info</summary>

```{r sessionInfo}
## DO NOT REMOVE!
## We store the settings of your computer and the current versions of the
## packages used to allow for reproducibility
Sys.time()
#git2r::repository() ## uncomment if you are using GitHub
sessionInfo()
```

</details>
