==> R/08_PVA_GLM.Rmd <==
---
title: "PVA of the NBI: GLM of survival and reproduction probabilities" ## name of your project
author: "Sinah Drenske, Viktoriia Radchuk, Stephanie Kramer-Schadt"                     ## your name(s)
date: "`r Sys.Date()`"                  ## current date
output:
  rmdformats::readthedown:
    code_folding: hide                  ## hide or show code by default?
    toc_depth: 3                        ## 3-level TOC
---

```{r setup, include=FALSE}
## You can ignore this chunk in most cases
## If you want to modify chunk options, you can do it here for all chunks or
## add the options in the repsective chunk header, e.g. `{r, message = FALSE}`
knitr::opts_chunk$set(echo = TRUE, warning = TRUE, message = TRUE,
                      fig.width = 9, fig.height = 6, dpi = 500, 
                      retina = 1, fig.showtext = TRUE)
```

* **Hypothesis:** We hypothesize that at the present time the northern bald ibis population can survive without further management and release. We predict that the observed demographic rates will ensure population growth and do not differ between the breeding colonies. 
* **Study area:** Austria, Germany and Italy
* **Data:** Data of the mean per scenario for the first NetLogo model

In this script, we created GAMs and GLMs to find out whether survival per stage or reproduction probabilities have a significant effect on lambda.

# Setup
```{r packages}
## for non-CRAN packages please keep install instruction
## but commented so it is not run each time, e.g.
# devtools::install_github("EcoDynIZW/template")

## libraries used in this script
## please add ALL LIBRARIES NEEDED HERE
## please remove libraries from the list that are not needed anymore 
## at a later stage
library("effects")
library("gam")
library("ggeffects")
library("here")
library("lattice")
library("mgcv")
library("modelsummary")
library('MuMIn')
library("plot3D")
library("pwr")
library("tidyverse")
```

# Data
```{r data}
df_sc_mean <- readRDS(file = here("output", "data-proc", "model-proc", "df_sc_mean.rds"))
```

# Management improvement scenarios
```{r}
unique(df_sc_mean$Repro)
```

Change the column names
```{r}
colnames(df_sc_mean)[10:14] <- c("m1", "m2", "m3", "m4", "RR") 
```

Add columns for survival
```{r}
df_sc_mean$s1 <- (1 - df_sc_mean$m1)
df_sc_mean$s2 <- (1 - df_sc_mean$m2)
df_sc_mean$s3 <- (1 - df_sc_mean$m3)
df_sc_mean$s4 <- (1 - df_sc_mean$m4)
```

Mortality and respective survival
```{r}
print("m1 vs s1")
table(df_sc_mean$m1)
table(df_sc_mean$s1)
print("m2 vs s2")
table(df_sc_mean$m2)
table(df_sc_mean$s2)
print("m3 vs s3")
table(df_sc_mean$m3)
table(df_sc_mean$s3)
print("m4 vs s4")
table(df_sc_mean$m4)
table(df_sc_mean$s4)
```

## Calculate extinction probability
```{r}
df_sc_mean$pext <- df_sc_mean$Extinct/100
```

## Explore database
```{r}
names(df_sc_mean)
View(df_sc_mean) # okay, these are mortalities.

unique(df_sc_mean[,c(20:24)]) # m1, m2, m3, m4, RR values per scenario
unique(df_sc_mean$s1)
unique(df_sc_mean$s2)
unique(df_sc_mean$s3)
unique(df_sc_mean$s4)
unique(df_sc_mean$RR)
```

```{r}
# start with pExt
table(df_sc_mean$pext)#df_sc_StevS_mean
#  0    0.01 0.02 0.03 0.06 0.24 
# 316    4    3    1    1    1 

# Comment: Only in 10 out of 326 scenarios did the population ever go extinct. 
# With such a distribution of the response variable, 
# there is no need to calculate a model. 
# Therefore we have described the distribution of survival and reproduction values

rows_ext_scen <- which(df_sc_mean$pext > 0)
ext_scen <- df_sc_mean[rows_ext_scen,]
ext_scen

table(ext_scen$m1)
table(ext_scen$m2)
table(ext_scen$m3)
table(ext_scen$m4)
table(ext_scen$RR)
# We can see that the highest mortality of adults was part of all scenarios where the population went extinct
# See also the excel file under "output\data-proc\model-proc\08_PVA_GLM_table_ext_scen.xlsx"
```

```{r}
# Explore lambda for s1 and RR
x <- df_sc_mean$s1
y <- df_sc_mean$RR
z <- df_sc_mean$Lambda_sto
scatter3D(x, y, z, pch = 19, cex = 0.7, clab = c("Lambda"), xlab = "s1", ylab = "RR", zlab = "Lambda")


contourplot(Lambda_sto ~ s1 * RR, data = df_sc_mean, labels = FALSE,
            region = TRUE)


plot(df_sc_mean$Lambda_sto)
abline(h=1) # Lambda > 1 indicates population growth

```
```{r}
# Explore lambda for s4 and RR
x <- df_sc_mean$s4
y <- df_sc_mean$RR
z <- df_sc_mean$Lambda_sto
scatter3D(x, y, z, pch = 19, cex = 0.7, clab = c("Lambda"), xlab = "s4", ylab = "RR", zlab = "Lambda")


contourplot(Lambda_sto ~ s4 * RR, data = df_sc_mean, labels = FALSE,
            region = TRUE)
```

# GAM
A general additive model (GAM) tests for non-linearity

The extinction probability is the response variable and the predictor variables are formed by the survival probabilities per stage (s1-s4) and the reproduction (RR)
```{r}
# define a k
# myk - stands for 'my number of knots/ smoothing splines' :-)
myk <- 3 # or 4
```

## Response = Lambda
```{r}
## Lambda and survival
gammod_2_1 <- gam(formula = Lambda_sto ~ s(s1, k=myk) + s(s2, k=myk) + 
                          s(s3, k=myk) + s(s4, k=myk) +  
                          s(RR, k=myk), 
                          data = df_sc_mean, 
                          na.action = 'na.fail', family=Gamma) 
summary(gammod_2_1) 
# F values for s4 and RR are very high - they have the highest influence

## Lambda and mortality
gammod_2_2 <- gam(formula = Lambda_sto ~ s(m1, k=myk) + s(m2, k=myk) + 
                          s(m3, k=myk) + s(m4, k=myk) +  
                          s(RR, k=myk), 
                          data = df_sc_mean, 
                          na.action = 'na.fail', family=Gamma) 
summary(gammod_2_2) 
# F values for s4 and RR are very high - they have the highest influence

# no difference between survival and mortality
```

## Plots
```{r}
plot(gammod_2_1)

vis.gam(gammod_2_1, view = c("s4", "RR"), plot.type = "persp", type = "response", zlim = c(0.5,1.8), 
        ticktype = "detailed", border = NA, n.grid = 1000, nticks = 3, 
        xlab = "s4", ylab = "RR", zlab = "Lambda", 
        color = "heat", theta= -140, cex.lab = 3, cex.axis = 2.5)

vis.gam(gammod_2_2, view = c("m4", "RR"), plot.type = "persp", type = "response", zlim = c(0.5,1.8), 
        ticktype = "detailed", border = NA, n.grid = 1000, nticks = 3, 
        xlab = "m4", ylab = "RR", zlab = "Lambda", 
        color = "heat", theta= -140, cex.lab = 3, cex.axis = 2.5)
```

Save the plot
```{r}
# s4 and RR
png(filename = here("plots", "04_PVA", "persp_s4_RR_lambda_20211124.png"), width = 1200, height = 772)
vis.gam(gammod_2_1, view = c("s4", "RR"), plot.type = "persp", type = "response", zlim = c(0.5,1.8),
        ticktype = "detailed", border = NA, n.grid = 1000, nticks = 3,
        xlab = "s4", ylab = "RR", zlab = "Lambda",
        color = "heat", theta= -140, cex.lab = 3, cex.axis = 2.5)
dev.off()
```

## Effect size
Calculating the effect size
```{r}
pwr.t.test(n=326, sig.level=0.05, power = 0.8, alternative="greater")
```


# GLM

## Response = Lambda
```{r}
# With interactions between RR and all survival values
mod1_3_1 <- glm(Lambda_sto ~ (s1 + s2 + s3 + s4) + RR, 
              data = df_sc_mean, na.action = 'na.fail', family=Gamma)

# With interactions between RR and all survival values
mod1_3_2 <- glm(Lambda_sto ~ (m1 + m2 + m3 + m4) + RR, 
              data = df_sc_mean, na.action = 'na.fail', family=Gamma)
```


```{r}
summary(mod1_3_1)  
anova(mod1_3_1, test='LRT') 
# s4 and RR have the highest deviance - they explain the model the most

#dredge(mod1_3_1)
plot(allEffects(mod1_3_1),type='response', ylim = c(1,2))
```

Save the plot
```{r}
png(filename = here("plots", "04_PVA", "lambda_effects_plot20211216.png"), width = 1200, height = 772)
plot(allEffects(mod1_3_1),type='response', ylim = c(1,2))
dev.off()
```


R square 
R^2= 1-(Residual deviance/Null deviance) 
```{r}
round(1 - (0.095412/1.507876), digits = 2) 
```

```{r}
summary(mod1_3_2) 
anova(mod1_3_2, test='LRT') 
# s4 and RR have the highest deviance - they explain the model the most

#dredge(mod1_3_2)
plot(allEffects(mod1_3_2),type='response')
```


R square 
R^2= 1-(Residual deviance/Null deviance) 
```{r}
round(1 - (0.095412/1.507876), digits = 2) 
```

## Model summary
```{r}
datasummary_skim(df_sc_mean)
varnam <- c('s1' = 'Survival of stage 1',
            's2' = 'Survival of stage 2',
            's3' = 'Survival of stage 3',
            's4' = 'Survival of stage 4',
            'RR' = 'Reproductive Rate')
varnam2 <- c('m1' = 'Mortality of stage 1',
            'm2' = 'Mortality of stage 2',
            'm3' = 'Mortality of stage 3',
            'm4' = 'Mortality of stage 4',
            'RR' = 'Reproductive Rate')


# modelplot(mod1_3,
#           coef_omit = 'Interc',
#           coef_map = varnam,
#           conf_level = .99) 

modelplot(mod1_3_1, 
          coef_omit = 'Interc',
          coef_map = varnam, 
          conf_level = .99) 

modelplot(mod1_3_2, 
          coef_omit = 'Interc',
          coef_map = varnam2, 
          conf_level = .99) 

# You can see that s4/m4 and RR are important and s1-s3/m1-m3 are not.
```




***

<details><summary>Session Info</summary>

```{r sessionInfo}
## DO NOT REMOVE!
## We store the settings of your computer and the current versions of the
## packages used to allow for reproducibility
Sys.time()
#git2r::repository() ## uncomment if you are using GitHub
sessionInfo()
```

</details>
